automatically generate theme data from feathers (included in texture information file, so we only need to load one file, only shapes and aeon_desktop for now)

+++ themes/from_feathers/get_themes.py
@@ -3,6 +3,7 @@ import shutil
 import argparse
 import parse_feathers_theme
 import xml_to_json
+import json
 
 # themes folder inside feathers folder
 FEATHERS_THEMES = 'themes'
@@ -33,7 +34,7 @@ def get_texture_name(theme_path):
             return texture_file.split('/')[-1]
 
 
-def get_themes(feathers_themes_path):
+def get_feathers_themes(feathers_themes_path):
     theme_data = {}
     themes = os.listdir(feathers_themes_path)
     
@@ -58,23 +59,41 @@ def get_themes(feathers_themes_path):
         theme_data[texture_name] = {
             'theme_path': theme_path,
             'base_file': base_path,
-            'texture_name': texture_name
+            'texture_name': texture_name,
+            'theme_texture': os.path.join(theme_path, 'assets', 'images', texture_name)
         }
-        print('found theme: {} ({})'.format(theme, texture_name))
+        print('found feathers theme: {} ({})'.format(theme, texture_name))
         
     return theme_data
 
 
-def process_assets(assets_path, name, theme_path):
-    # copy assets from feathers and convert .xml to .json
-    theme_texture = os.path.join(theme_path, 'assets', 'images', name)
+def create_theme_dir(assets_path, name):
     gown_asset = os.path.join(assets_path, name)+os.path.sep
     if not os.path.exists(gown_asset):
         os.makedirs(gown_asset)
-    shutil.copy2(theme_texture+'.png', gown_asset)
-    shutil.copy2(theme_texture+'.xml', gown_asset)
+    return gown_asset
 
-    xml_to_json.convert(os.path.join(gown_asset, name+'.xml'))
+
+def process_assets(assets_path, name, theme_path):
+    gown_asset = create_theme_dir(assets_path, name)
+    # copy assets from feathers and convert .xml to .json
+    theme_texture = os.path.join(theme_path, 'assets', 'images', name)
+    create_theme_dir(assets_path, name)
+    if os.path.exists(theme_texture+'.png'):
+        shutil.copy2(theme_texture+'.png', gown_asset)
+
+
+def get_theme_data(data_path):
+    data = {}
+    for name in os.listdir(data_path):
+        name, ending = os.path.splitext(name)
+        if ending != '.json':
+            continue
+        print('found theme data: {}'.format(name))
+        data.setdefault(name, {})
+        data[name]['data_path'] = os.path.join(data_path, name+'.json') #json.load(file(path, 'r'))
+    return data
+
 
 def main():
     parser = argparse.ArgumentParser(
@@ -83,17 +102,43 @@ def main():
                         help='path to feathers directory')
     parser.add_argument('assets_path', type=str,
                         help='Path to GOWN assets where we extract the theme to')
+    parser.add_argument('data_path', type=str,
+                        help='Path to manual theme data that will be merged with texture data')
     args = parser.parse_args()
-    theme_data = get_themes(os.path.join(args.feathers_path, FEATHERS_THEMES))
+    theme_data = get_theme_data(args.data_path)
+    
+    feathers_data = get_feathers_themes(os.path.join(args.feathers_path, FEATHERS_THEMES))
+    for name, theme in feathers_data.items():
+        theme_data.setdefault(name, {})
+        theme_data[name].update(theme)
     
     for name, theme in theme_data.items():
-        # copy asset files to new folder
-        process_assets(args.assets_path, name, theme['theme_path'])
+        data = {}
+        
+        # get texture positions and sizes from atlas file
+        if 'theme_texture' in theme:
+            data.update(xml_to_json.get_data(theme['theme_texture']+'.xml'))
         
         # get scale 9 values from Base*.as file
-        parse_feathers_theme.convert(theme['base_file'], 
-                os.path.join(args.assets_path, name, name))
+        if 'base_file' in theme:
+            grids = parse_feathers_theme.get_scale_9_grids(theme['base_file'])
+            if grids:
+                data['grids'] = grids
         
+        # manual entered data (as last step so the user can overwrite everything)
+        if 'data_path' in theme:
+            data.update(json.load(file(theme['data_path'], 'r')))
+
+        # now that we have the data we just need to store everything in 
+        # the assets-folder and copy the image
+        create_theme_dir(args.assets_path, name)
+        if 'theme_path' in theme:
+            process_assets(args.assets_path, name, theme['theme_path'])
+            
+        json_path = os.path.join(args.assets_path, name, name+'.json')
+        json.dump(data, file(json_path, 'w'), indent=2)
+
+            
 
 
 if __name__ == '__main__':
